+++ Packages/Accelib/Module/Accelib.Preview/IPreviewIconProvider.cs
@@ -1,9 +1,0 @@
-﻿using Sirenix.OdinInspector;
-
-namespace Accelib.Preview
-{
-    public interface IPreviewIconProvider
-    {
-        public SdfIconType EditorPreviewIcon { get; }
-    }
-}