+++ Packages/Accelib/Module/Accelib.Preview/IPreviewNameProvider.cs
@@ -1,7 +1,0 @@
-﻿namespace Accelib.Preview
-{
-    public interface IPreviewNameProvider
-    {
-        public string EditorPreviewName { get; }
-    }
-}