+++ Packages/Accelib/Module/Accelib.Preview/Runtime/ISubAssetProvider.cs
@@ -1,8 +1,9 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace Accelib.Preview
 {
+    /// <summary>하위 ScriptableObject 에셋 목록을 제공하는 인터페이스.</summary>
     public interface ISubAssetProvider
     {
         public IReadOnlyList<ScriptableObject> SubAssets { get; }